+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
  - pipe
+ - splitAt/asyncSplitAt
 
 ### Changes
  - takeSorted/asyncTakeSorted optional comparator is now the first argument